Revisado tipo de datos de los modelos y pruebas tarifas

+++ lib/src/features/articulos/domain/articulo_recambio.dart
@@ -9,7 +9,7 @@ class ArticuloRecambio with _$ArticuloRecambio {
       {required String articuloId,
       required String id,
       required String descripcion,
-      required double cantidad,
+      required int cantidad,
       required DateTime lastUpdated,
       required bool deleted}) = _ArticuloRecambio;
 }
